% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ensemble.R
\name{add_ensemble}
\alias{add_ensemble}
\title{Add an Ensemble Model of multiple Post-Processing Methods}
\usage{
add_ensemble(
  df_combined,
  train_val_split = TRUE,
  verbose = FALSE,
  max_iter = 1e+05,
  print_level = 0
)
}
\description{
Add an Ensemble Model of multiple Post-Processing Methods
}
