% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_predictions.R
\name{collect_predictions}
\alias{collect_predictions}
\title{Collect Updated Dataframes for each Post Processing Method}
\usage{
collect_predictions(...)
}
\description{
Collect Updated Dataframes for each Post Processing Method
}
