% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_methods.R
\name{eval_methods}
\alias{eval_methods}
\title{Compute the relative/percentage changes in the Weighted Interval Score of multiple
Post-Processing Methods compared to the original Forecasts}
\usage{
eval_methods(
  df_combined,
  summarise_by,
  training_set = FALSE,
  margins = FALSE,
  row_averages = FALSE,
  col_averages = FALSE,
  round_digits = 4
)
}
\description{
Compute the relative/percentage changes in the Weighted Interval Score of multiple
Post-Processing Methods compared to the original Forecasts
}
