% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sets.R
\name{extract_validation_set}
\alias{extract_validation_set}
\title{Extract all Observations that were contained in the Validation Set during Model Fitting}
\usage{
extract_validation_set(df_combined, cv_init_training = NULL)
}
\description{
Extract all Observations that were contained in the Validation Set during Model Fitting
}
