% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_eval}
\alias{plot_eval}
\title{Visualize the Output of \code{eval_methods()} as a Barplot or Heatmap}
\usage{
plot_eval(df_eval, heatmap = TRUE, base_size = 9)
}
\description{
Visualize the Output of \code{eval_methods()} as a Barplot or Heatmap
}
