% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_intervals}
\alias{plot_intervals}
\title{Visualize updated Forecast Intervals for multiple Post-Processing Methods}
\usage{
plot_intervals(
  df,
  model = NULL,
  location = NULL,
  target_type = c("Cases", "Deaths"),
  quantile = 0.05,
  horizon = 1,
  highlight_cv = TRUE,
  highlight_time_point = NULL,
  base_size = 9
)
}
\description{
Visualize updated Forecast Intervals for multiple Post-Processing Methods
}
