% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_intervals_grid}
\alias{plot_intervals_grid}
\title{Visualize updated Forecast Intervals along the \code{target_type} and \code{horizon} or
\code{quantile} dimensions}
\usage{
plot_intervals_grid(
  df,
  model = NULL,
  location = NULL,
  facet_by = c("horizon", "quantile"),
  quantiles = NULL,
  horizon = NULL,
  highlight_cv = FALSE,
  base_size = 9
)
}
\description{
Visualize updated Forecast Intervals along the \code{target_type} and \code{horizon} or
\code{quantile} dimensions
}
