% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_predictions.R
\name{update_predictions}
\alias{update_predictions}
\title{Update Original Forecasts with multiple Post Processing Methods}
\usage{
update_predictions(
  df,
  methods = c("cqr", "cqr_asymmetric", "cqr_multiplicative", "qsa_uniform",
    "qsa_flexible", "qsa_flexible_symmetric"),
  models = NULL,
  locations = NULL,
  target_types = NULL,
  horizons = NULL,
  quantiles = NULL,
  cv_init_training = NULL,
  penalty_weight = NULL,
  optim_method = NULL,
  lower_bound_optim = 0,
  upper_bound_optim = 5,
  steps_optim = 0.1,
  parallel = FALSE,
  regularize_scores = FALSE,
  constrain_margins = FALSE,
  return_list = TRUE,
  verbose = FALSE
)
}
\description{
Update Original Forecasts with multiple Post Processing Methods
}
